BZ2-2328 fixed ts linter
