+++ packages/Core/src/ThemeContext.tsx
@@ -1,13 +1,13 @@
 import React from 'react';
 
 
-const ThemeContext = React.createContext({ 
-  dark:{
+const ThemeContext = React.createContext({
+  dark: {
   },
-  default:{
-    button:''
+  default: {
+    button: ''
   },
-  light:{
+  light: {
   },
   selectedTheme: 'default',
 });
