+++ packages/Core/src/interfaces.tsx
@@ -1,0 +1,12 @@
+interface IThemeStructure {
+  button: string;
+}
+
+interface IThemeContext {
+  dark:  object;
+  default:  IThemeStructure;
+  light:  object;
+  selectedTheme: string;
+}
+
+export { IThemeStructure, IThemeContext }
