+++ packages/Core/src/useTheme.tsx
@@ -1,21 +1,11 @@
 import { useContext } from "react";
+import { IThemeContext } from './interfaces'
 import { ThemeContext } from './ThemeContext'
-
-interface IThemeStructure {
-  button: string;
-}
-
-interface IThemeContext {
-  dark:  object;
-  default:  IThemeStructure;
-  light:  object;
-  selectedTheme: string;
-}
 
 const useTheme = (componentName: string):string => {
    const themeOptions = useContext<IThemeContext>(ThemeContext);
-   const {selectedTheme} = themeOptions;
-   const componentThemeClasses = themeOptions[selectedTheme][componentName] || '';
+   const { selectedTheme } = themeOptions;
+    const componentThemeClasses = themeOptions[selectedTheme][componentName] || '';
 
   return componentThemeClasses;
 }
