+++ packages/Group/package.json
@@ -1,0 +1,36 @@
+{
+  "name": "@blaze-react/group",
+  "version": "0.6.0",
+  "publishConfig": {
+    "access": "public"
+  },
+  "description": "",
+  "main": "./lib/index.js",
+  "typings": "./lib/index.d.ts",
+  "scripts": {
+    "build": "../../node_modules/.bin/rimraf ./lib && ../../node_modules/.bin/tsc --jsx react --esModuleInterop ./src/* --outDir lib ",
+    "lint": "../../node_modules/.bin/tslint -c ../../tslint.json src/**/*.tsx",
+    "test": "jest",
+    "test:coverage": "jest --coverage",
+    "pre-publish": "npm run build && npm run lint && npm run test"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "devDependencies": {
+    "@babel/cli": "^7.4.4",
+    "@babel/core": "^7.4.5"
+  },
+  "peerDependencies": {
+    "prop-types": "^15.7.2",
+    "react": "^16.9.0",
+    "react-dom": "^16.9.0"
+  },
+  "babel": {
+    "presets": [
+      "@babel/preset-env",
+      "@babel/preset-react"
+    ]
+  },
+  "gitHead": "c7cc1d26b4046b800c727feb1dd283dbd833328c"
+}
