+++ packages/Group/src/interfaces.tsx
@@ -1,0 +1,5 @@
+// interface IGroupProps {
+
+// }
+
+// export { IGroupProps };
